+++ example/lib/count_up_timer_page.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:stop_watch_timer/stop_watch_timer.dart';
-import 'package:stop_watch_timer_example/rounded_button.dart';
 
 class CountUpTimerPage extends StatefulWidget {
   static Future<void> navigatorPush(BuildContext context) async {
@@ -60,7 +59,7 @@ class _State extends State<CountUpTimerPage> {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: const Text('CountUp Sample'),
+        title: const Text('Count Up Timer'),
       ),
       body: Scrollbar(
         child: SingleChildScrollView(
@@ -249,36 +248,36 @@ class _State extends State<CountUpTimerPage> {
                 Row(
                   mainAxisAlignment: MainAxisAlignment.center,
                   children: <Widget>[
-                    Padding(
-                      padding: const EdgeInsets.symmetric(horizontal: 4),
-                      child: RoundedButton(
-                        color: Colors.lightBlue,
-                        onTap: _stopWatchTimer.onStartTimer,
-                        child: const Text(
-                          'Start',
-                          style: TextStyle(color: Colors.white),
+                    Flexible(
+                      child: Padding(
+                        padding: const EdgeInsets.symmetric(horizontal: 4),
+                        child: FilledButton(
+                          onPressed: _stopWatchTimer.onStartTimer,
+                          child: const Text(
+                            'Start',
+                          ),
                         ),
                       ),
                     ),
-                    Padding(
-                      padding: const EdgeInsets.symmetric(horizontal: 4),
-                      child: RoundedButton(
-                        color: Colors.green,
-                        onTap: _stopWatchTimer.onStopTimer,
-                        child: const Text(
-                          'Stop',
-                          style: TextStyle(color: Colors.white),
+                    Flexible(
+                      child: Padding(
+                        padding: const EdgeInsets.symmetric(horizontal: 4),
+                        child: FilledButton(
+                          onPressed: _stopWatchTimer.onStopTimer,
+                          child: const Text(
+                            'Stop',
+                          ),
                         ),
                       ),
                     ),
-                    Padding(
-                      padding: const EdgeInsets.symmetric(horizontal: 4),
-                      child: RoundedButton(
-                        color: Colors.red,
-                        onTap: _stopWatchTimer.onResetTimer,
-                        child: const Text(
-                          'Reset',
-                          style: TextStyle(color: Colors.white),
+                    Flexible(
+                      child: Padding(
+                        padding: const EdgeInsets.symmetric(horizontal: 4),
+                        child: FilledButton(
+                          onPressed: _stopWatchTimer.onResetTimer,
+                          child: const Text(
+                            'Reset',
+                          ),
                         ),
                       ),
                     ),
@@ -289,14 +288,14 @@ class _State extends State<CountUpTimerPage> {
                   child: Row(
                     mainAxisAlignment: MainAxisAlignment.center,
                     children: <Widget>[
-                      Padding(
-                        padding: const EdgeInsets.all(0).copyWith(right: 8),
-                        child: RoundedButton(
-                          color: Colors.deepPurpleAccent,
-                          onTap: _stopWatchTimer.onAddLap,
-                          child: const Text(
-                            'Lap',
-                            style: TextStyle(color: Colors.white),
+                      Flexible(
+                        child: Padding(
+                          padding: const EdgeInsets.all(0).copyWith(right: 8),
+                          child: FilledButton(
+                            onPressed: _stopWatchTimer.onAddLap,
+                            child: const Text(
+                              'Lap',
+                            ),
                           ),
                         ),
                       ),
@@ -308,69 +307,72 @@ class _State extends State<CountUpTimerPage> {
                   child: Row(
                     mainAxisAlignment: MainAxisAlignment.center,
                     children: <Widget>[
-                      Padding(
-                        padding: const EdgeInsets.symmetric(horizontal: 4),
-                        child: RoundedButton(
-                          color: Colors.pinkAccent,
-                          onTap: () {
-                            _stopWatchTimer.setPresetHoursTime(1);
-                          },
-                          child: const Text(
-                            'Set Hours',
-                            style: TextStyle(color: Colors.white),
+                      Flexible(
+                        child: Padding(
+                          padding: const EdgeInsets.symmetric(horizontal: 4),
+                          child: FilledButton(
+                            onPressed: () {
+                              _stopWatchTimer.setPresetHoursTime(1);
+                            },
+                            child: const Text(
+                              'Set Hours',
+                            ),
                           ),
                         ),
                       ),
-                      Padding(
-                        padding: const EdgeInsets.symmetric(horizontal: 4),
-                        child: RoundedButton(
-                          color: Colors.pinkAccent,
-                          onTap: () {
-                            _stopWatchTimer.setPresetMinuteTime(59);
-                          },
-                          child: const Text(
-                            'Set Minute',
-                            style: TextStyle(color: Colors.white),
-                          ),
-                        ),
-                      ),
-                      Padding(
-                        padding: const EdgeInsets.symmetric(horizontal: 4),
-                        child: RoundedButton(
-                          color: Colors.pinkAccent,
-                          onTap: () {
-                            _stopWatchTimer.setPresetSecondTime(10);
-                          },
-                          child: const Text(
-                            'Set Second',
-                            style: TextStyle(color: Colors.white),
+                      Flexible(
+                        child: Padding(
+                          padding: const EdgeInsets.symmetric(horizontal: 4),
+                          child: FilledButton(
+                            onPressed: () {
+                              _stopWatchTimer.setPresetMinuteTime(59);
+                            },
+                            child: const Text(
+                              'Set Minute',
+                            ),
                           ),
                         ),
                       ),
                     ],
                   ),
                 ),
-                Padding(
-                  padding: const EdgeInsets.symmetric(horizontal: 4),
-                  child: RoundedButton(
-                    color: Colors.pinkAccent,
-                    onTap: () {
-                      _stopWatchTimer.setPresetTime(mSec: 3599 * 1000);
-                    },
-                    child: const Text(
-                      'Set PresetTime',
-                      style: TextStyle(color: Colors.white),
+                Row(
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  children: [
+                    Flexible(
+                      child: Padding(
+                        padding: const EdgeInsets.symmetric(horizontal: 4),
+                        child: FilledButton(
+                          onPressed: () {
+                            _stopWatchTimer.setPresetSecondTime(10);
+                          },
+                          child: const Text(
+                            'Set +Second',
+                          ),
+                        ),
+                      ),
                     ),
-                  ),
+                    Flexible(
+                      child: Padding(
+                        padding: const EdgeInsets.symmetric(horizontal: 4),
+                        child: FilledButton(
+                          onPressed: () {
+                            _stopWatchTimer.setPresetSecondTime(-10);
+                          },
+                          child: const Text(
+                            'Set -Second',
+                          ),
+                        ),
+                      ),
+                    ),
+                  ],
                 ),
                 Padding(
                   padding: const EdgeInsets.symmetric(horizontal: 4),
-                  child: RoundedButton(
-                    color: Colors.pinkAccent,
-                    onTap: _stopWatchTimer.clearPresetTime,
+                  child: FilledButton(
+                    onPressed: _stopWatchTimer.clearPresetTime,
                     child: const Text(
                       'Clear PresetTime',
-                      style: TextStyle(color: Colors.white),
                     ),
                   ),
                 ),